chore: update release scripts
Improve the commit message added to release commits. It now includes a
complete change log for the release being made.

When the `console-api` version changes, it also needs to be changed in
the crates that depend on it. The release script now does this via cargo
update-version (from cargo-edit).

Committing the `Cargo.lock` file after publishing has been moved to
before adding the version tag to ensure that the correct commit is
tagged. The change log is also updated in the amended commit so that the
hash is correct.

The cliff configuration file has also been updated to fix some
incorrectly formatted commits.