docs: update MSRV version docs to 1.64
PR #464 increased the MSRV to Rust 1.64, including the CI MSRV job and
the `rust-version` fields in `Cargo.toml`. However, it did not update
the MSRV documentation in README.md files for the console crates. This
commit updates the READMEs as well.

It would be really cool if we could define the MSRV in a single place
and have it control CI, the `Cargo.toml` `rust-version` field, *and*
docs, but that's probably a bunch of extra work. So, for now, I've
manually made sure everything is in sync.