chore(console): update Clap to v4.1 (#469)
Tokio console was previously using Clap v3, which is in maintenance and                             
doesn't always receive necessary updates. For example, the fix for a new                            
lint in Clippy was submitted by us (and updated in #414).                                           
                                                                                                    
This change updates to Clap 4.1, as the 4.x versions are generally                                  
better supported. This is the latest version of Clap which is compatible                            
with our current MSRV of 1.64. From the 4.2 versions there are some                                 
depdendencies which require Rust 1.70.                                                              
                                                                                                    
Additionally, we switched from using a `tracing_subscriber::EnvFilter`                              
to a `tracing_subscriber::filter::Targets` for filtering the console's                              
internal log messages. This is because `clap` requires that all arg                                 
values be `Clone` in v4, and we weren't using `EnvFilter` span filtering                            
anyway, so we felt like this was the simplest option.                                               
                                                                                                    
This change shouldn't result in any change in behavior besides                                      
differences in CLI help text formatting.

This change is based on the work originally done in #384.