fix(console): ignore key release events (#468)
On supported platforms, a short key press generates two crossterm
events: One for the initial press and another when the key is released
again. The code was not checking the event kind, so it would treat a
key release as a second key press. Filter out key releases entirely
since they're not explicitly used for anything.