fix(console): remove duplicate controls from async ops view (#519)
Each screen in Tokio Console has a controls section that lists all the
possible keys that can be used to control that screen. However, on the
Resource Detail screen, there were two controls sections - one at the
top of the screen and another above the Async Ops table.

Since the controls for the Async Ops view were already included in the
Resource Detail view controls panel in #427, the controls for the Async
Ops view only needed to be removed. This simplified the layout somewhat,
as the Async Ops view now only contains a single-subview, the table
itself.

Closes #258