chore: automate releases with Release-plz
Releasing the console crates has been a bit of a hassle. It requires
manual work run locally and as such, we have gone long periods without
releases, despite having changes which would be worth releasing.

[Release-plz] provides release automation based on GitHub actions. It
will create, and then update a PR for the next release.

Release-plz depends on [conventional commits] (which the `console`
project is already following) to determine the next version number for
each crate and then uses [cargo-semver-checks] to check that the version
bump is correct.

For the changelogs, [git-cliff] is used. Since we already started using
in #416, so we can reuse our existing `cliff.toml` config file. A small
change has been made to remove the footer, as it was getting included
multiple times in a single changelog file due to the delta updates.

**Caution**: In order to use Release-plz, a crates.io token first needs
to be added to the GitHub project. This means that anyone with PR merge
rights will be able to run a release by merging the Release PR.
Additionally, GitHub actions need to be given permission to create PRs.

The previous scripts used to release have been removed.

[Release-plz]: https://release-plz.ieni.dev/
[conventional commits]: https://www.conventionalcommits.org/en/v1.0.0/
[cargo-semver-checks]: https://github.com/obi1kenobi/cargo-semver-checks
[git-cliff]: https://git-cliff.org/