chore: get rid of remove_dir_all (#542)
The remove_dir_all crate is a Rust library that offers additional 
features over the Rust standard library fs::remove_dir_all function. 
It suffers the same class of failure as the code it was layering over:
TOCTOU race conditions, with the ability to cause arbitrary paths 
to be deleted by substituting a symlink for a path after the type of 
the path was checked. 

So we need to get rid of the remove_dir_all dependency.