feat(subscriber): add `TOKIO_CONSOLE_BUFFER_CAPACITY`  env variable (#568)
When the recorded process spawns a lot of tasks and execute a lot of
operations, a red message can be seen on the top of the console saying
`dropped: 14319181 async_ops, 1601630 tasks, 634586 resources`. That
means that the buffer is not sufficiently big to record these and
someone in the tokio Discord advised to use the
`TOKIO_CONSOLE_BUFFER_CAPACITY` variable to tune that buffer. Turns out
it was not a thing, so this pr adds it.