update: bump tonic to 0.12
Update all dependencies on `tonic` and `tonic-build` to v0.12, along
with the related `prost`, `prost-types` and `prost-build` crates to
v0.13.1.

BREAKING CHANGE:
This is a breaking change for users of `console-api` and
`console-subscriber`, as it changes the public `tonic`, `prost` and
`prost-types` dependency to a semver-incompatible version. This breaks
compatibility with `tonic` 0.11.x as well as `prost`/`prost-types`
0.12.x.