chore: update tokio-console screenshots for v0.1.13
The screenshots on the docs.rs page for tokio-console are a bit out of
date. They are from v0.1.8 and still show the Target column instead of
the Kind column.

With the task size instrumentation and new lints which will be released
in the next version of Tokio Console, we need screenshots which match
what the user will see. The same goes for the screenshots in the
repository root README.md, which anyone visiting the repo page on GitHub
will see.

This change adds new screenshots for docs.rs and replaces the README
screenshots. We have to keep the docs.rs images from previous versions
still, since docs.rs pages for previous versions still reference them.

Additionally, a new CI job to build the docs.rs documentation and also
to check that all the images referenced from the tokio-console README
file (which is used on docs.rs) are present in the repository, as we
can't inspect the documentation until the change is merged.