fix(console): add dynamic constraints layout in task details screen (#614)
The changes I've made is for accommodating long location names.

I've changed the .. to the long name; it widens the task rectangle if its longer
than the default (50%) and goes to the next line if its longer than what we can
accommodate in a single line. This fixes issue #523, I have attached screenshots
of the same in the PR, before and after changes.

Fixes #523

Co-authored-by: Hayden Stainsby <hds@caffeineconcepts.com>